+++ CHANGELOG.md
@@ -9,11 +9,13 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- ilivalidator configuration file (aka `toml`) support. One configuration file for one INTERLIS model only (baked into resources directory). It will be used by default if one is present. It looks for a *.toml file with the same name as the data model name. Config file name must be lower-case.
+- ilivalidator configuration file (aka `toml`) support. One configuration file for one INTERLIS model only (baked into resources directory). It will be used by default if one is present. It looks for a *.toml file with the same name as the data model name. The configuration file name must be lower-case.
 - Added and adjusted some scripts for AWS Codepipeline with deployment on EBS.
 
 ### Changed
 
+- Spring Boot 1.5.6
 - ilivalidator-1.4.0
+- Use sogeo.services artifactory repository instead of static maven repository.
 - Use `java.io.tmpdir` as default temporary directory. Can be overriden with the `ch.so.agi.ilivalidator.uploadedFiles` in the `application.properties` file.